feat(goesr): rename netcdf_image module to image
