+++ src/goesdl/goesr/image.py
